# frontend/src/app/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "SQL Server Performance Lab API",
    "version": "v1"
  },
  "paths": {
    "/api/v1/labs": {
      "get": {
        "tags": [
          "Labs"
        ],
        "operationId": "ListLabs",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LabDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Labs"
        ],
        "operationId": "CreateLab",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLabRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateLabResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/labs/{labId}": {
      "get": {
        "tags": [
          "Labs"
        ],
        "operationId": "GetLab",
        "parameters": [
          {
            "name": "labId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Labs"
        ],
        "operationId": "DeleteLab",
        "parameters": [
          {
            "name": "labId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteLabRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/labs/{labId}/start": {
      "post": {
        "tags": [
          "Labs"
        ],
        "operationId": "StartLab",
        "parameters": [
          {
            "name": "labId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/labs/{labId}/deallocate": {
      "post": {
        "tags": [
          "Labs"
        ],
        "operationId": "DeallocateLab",
        "parameters": [
          {
            "name": "labId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/labs/{labId}/extend-expiration": {
      "post": {
        "tags": [
          "Labs"
        ],
        "operationId": "ExtendLabExpiration",
        "parameters": [
          {
            "name": "labId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtendExpirationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/labs/{labId}/jobs": {
      "get": {
        "tags": [
          "Labs"
        ],
        "operationId": "ListLabJobs",
        "parameters": [
          {
            "name": "labId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/labs/{labId}/audit": {
      "get": {
        "tags": [
          "Labs"
        ],
        "operationId": "ListLabAudit",
        "parameters": [
          {
            "name": "labId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEventDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{jobId}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "operationId": "GetJob",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{jobId}/cancel": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "operationId": "CancelJob",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/environment": {
      "get": {
        "tags": [
          "Environment"
        ],
        "operationId": "GetEnvironment",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvironmentDto"
                }
              }
            }
          }
        }
      }
    },
    "/config.json": {
      "get": {
        "tags": [
          "Environment"
        ],
        "operationId": "GetClientConfig",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientConfigDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuditEventDto": {
        "required": [
          "id",
          "labId",
          "jobId",
          "actorId",
          "action",
          "detail",
          "correlationId",
          "occurredAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "labId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "jobId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "actorId": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "correlationId": {
            "type": "string"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ClientConfigDto": {
        "required": [
          "authMode",
          "tenantId",
          "clientId",
          "authority",
          "apiScope"
        ],
        "type": "object",
        "properties": {
          "authMode": {
            "type": "string"
          },
          "tenantId": {
            "type": [
              "null",
              "string"
            ]
          },
          "clientId": {
            "type": [
              "null",
              "string"
            ]
          },
          "authority": {
            "type": [
              "null",
              "string"
            ]
          },
          "apiScope": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "CreateLabRequest": {
        "required": [
          "name",
          "region",
          "ttlHours"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "ttlHours": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CreateLabResponse": {
        "required": [
          "lab",
          "job"
        ],
        "type": "object",
        "properties": {
          "lab": {
            "$ref": "#/components/schemas/LabDto"
          },
          "job": {
            "$ref": "#/components/schemas/JobDto"
          }
        }
      },
      "DeleteLabRequest": {
        "required": [
          "confirmName"
        ],
        "type": "object",
        "properties": {
          "confirmName": {
            "type": "string"
          }
        }
      },
      "EnvironmentDto": {
        "required": [
          "infrastructureMode",
          "isSimulated",
          "authenticationMode",
          "regions",
          "minTtlHours",
          "maxTtlHours",
          "maxExtensionHours",
          "maxActiveLabsPerUser"
        ],
        "type": "object",
        "properties": {
          "infrastructureMode": {
            "type": "string"
          },
          "isSimulated": {
            "type": "boolean"
          },
          "authenticationMode": {
            "type": "string"
          },
          "regions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "minTtlHours": {
            "type": "integer",
            "format": "int32"
          },
          "maxTtlHours": {
            "type": "integer",
            "format": "int32"
          },
          "maxExtensionHours": {
            "type": "integer",
            "format": "int32"
          },
          "maxActiveLabsPerUser": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ErrorCategory": {
        "enum": [
          "Transient",
          "Permanent",
          "Authorization",
          "Quota",
          "Validation",
          "Cancellation",
          null
        ]
      },
      "ExtendExpirationRequest": {
        "required": [
          "hours"
        ],
        "type": "object",
        "properties": {
          "hours": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "JobDto": {
        "required": [
          "jobId",
          "labId",
          "type",
          "status",
          "progress",
          "currentStep",
          "attemptCount",
          "maxAttempts",
          "requestedBy",
          "requestedAt",
          "startedAt",
          "completedAt",
          "updatedAt",
          "cancelRequested",
          "errorCategory",
          "errorCode",
          "errorMessage",
          "statusUrl"
        ],
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid"
          },
          "labId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "$ref": "#/components/schemas/LabJobType"
          },
          "status": {
            "$ref": "#/components/schemas/LabJobStatus"
          },
          "progress": {
            "type": "integer",
            "format": "int32"
          },
          "currentStep": {
            "type": [
              "null",
              "string"
            ]
          },
          "attemptCount": {
            "type": "integer",
            "format": "int32"
          },
          "maxAttempts": {
            "type": "integer",
            "format": "int32"
          },
          "requestedBy": {
            "type": "string"
          },
          "requestedAt": {
            "type": "string",
            "format": "date-time"
          },
          "startedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "completedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "cancelRequested": {
            "type": "boolean"
          },
          "errorCategory": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ErrorCategory"
              }
            ]
          },
          "errorCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "errorMessage": {
            "type": [
              "null",
              "string"
            ]
          },
          "statusUrl": {
            "type": "string"
          }
        }
      },
      "LabCommand": {
        "enum": [
          "Start",
          "Deallocate",
          "Delete",
          "ExtendExpiration"
        ]
      },
      "LabDto": {
        "required": [
          "id",
          "name",
          "ownerId",
          "ownerName",
          "region",
          "resourceGroupName",
          "state",
          "stateReason",
          "powerState",
          "isSimulated",
          "createdAt",
          "updatedAt",
          "expiresAt",
          "currentJob",
          "allowedActions"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "ownerName": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "resourceGroupName": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/LabState"
          },
          "stateReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "powerState": {
            "$ref": "#/components/schemas/VmPowerState"
          },
          "isSimulated": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "currentJob": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JobDto"
              }
            ]
          },
          "allowedActions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LabCommand"
            }
          }
        }
      },
      "LabJobStatus": {
        "enum": [
          "Queued",
          "Running",
          "Succeeded",
          "Failed",
          "Cancelled"
        ]
      },
      "LabJobType": {
        "enum": [
          "ProvisionLab",
          "StartVm",
          "DeallocateVm",
          "DeleteLab",
          "SeedDatabase",
          "ApplyIndex",
          "RemoveIndex",
          "RunIndexBenchmark",
          "RunDeadlock",
          "RunDeadlockResolution",
          "CollectMetrics",
          "CreateBackup",
          "VerifyBackup",
          "AssessPatches",
          "InstallPatches",
          "ReconcileLab"
        ]
      },
      "LabState": {
        "enum": [
          "Requested",
          "Provisioning",
          "Stopped",
          "Starting",
          "Configuring",
          "Ready",
          "Deallocating",
          "Maintenance",
          "Deleting",
          "Deleted",
          "Failed"
        ]
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "VmPowerState": {
        "enum": [
          "Unknown",
          "NotCreated",
          "Starting",
          "Running",
          "Deallocating",
          "Deallocated",
          "Deleted"
        ]
      }
    }
  },
  "tags": [
    {
      "name": "Labs"
    },
    {
      "name": "Jobs"
    },
    {
      "name": "Environment"
    }
  ]
}